combined code in add to cart and merch content
