+++ final-project/src/components/Merch-Content.jsx
@@ -1,37 +1,82 @@
 
-import React from "react";
-import axios from "axios";
-// import './styles/Merch-Content.css';
+// import React from "react";
+// import axios from "axios";
+// // import './styles/Merch-Content.css';
 
-// Fetch a random merch item
-const FetchMerch = async () => {
-    try {
-      const response = await axios.get('https://fakestoreapi.com/products/1')
-      .then(res=>res.json())
-      .then(json=>console.log(json))
+// // Fetch a random merch item
+// const FetchMerch = async () => {
+//     try {
+//       const response = await axios.get('https://fakestoreapi.com/products/1')
+//       .then(res=>res.json())
+//       .then(json=>console.log(json))
 
-    }
-   catch (error) {
-    console.error("Error fetching videos:", error);
-    setError("Failed to fetch videos.");
-  }
+//     }
+//    catch (error) {
+//     console.error("Error fetching videos:", error);
+//     setError("Failed to fetch videos.");
+//   }
 
-  return(
+//   return(
+//     <div className='product-display'>
+//       {products.map((product) => (
+//         <div key={product.id}>
+//           <h2>{product.title}</h2>
+//           <p>${product.price}</p>
+//           <p>{product.image}</p>
+//           <button onClick={() => addToCart(product)}>Add to Cart</button>
+//         </div>
+//       ))}
+//     </div>
+//   )
+// };
+
+// export default FetchMerch; 
+
+
+import React, { useState, useEffect } from 'react';
+import axios from 'axios';
+
+
+const FetchProducts = () => {
+  const [products, setProducts] = useState([]);
+  const [error, setError] = useState('');
+  
+  useEffect(() => {
+    const fetchProducts = async () => {
+      try {
+        
+        const response = await axios.get('https://fakestoreapi.com/products?limit=1');
+        const data = response.data;  
+        setProducts(data);
+      } catch (error) {
+        console.error("Error fetching products:", error);
+        setError("Failed to fetch products.");
+      }
+    };
+
+    fetchProducts();
+  }, []);
+
+  const addToCart = (product) => {
+    // Add "add to cart" logic here
+    console.log(`Added ${product.title} to cart`);
+  };
+
+  return (
     <div className='product-display'>
+      {error && <p>{error}</p>}
       {products.map((product) => (
         <div key={product.id}>
           <h2>{product.title}</h2>
+          <img src={product.image} alt={product.title} />
           <p>${product.price}</p>
-          <p>{product.image}</p>
           <button onClick={() => addToCart(product)}>Add to Cart</button>
         </div>
       ))}
     </div>
-  )
+  );
 };
 
-export default FetchMerch; 
-
-
+export default FetchProducts;
