+++ final-project/src/components/buttons/Add-To-Cart-Button.jsx
@@ -1,47 +1,47 @@
 
-import React, { useState, useEffect } from 'react';
-import '../styles/addToCart.css';
+// import React, { useState, useEffect } from 'react';
+// import '../styles/addToCart.css';
 
-const FetchProducts = () => {
-  const [products, setProducts] = useState([]);
-  const [error, setError] = useState('');
+// const FetchProducts = () => {
+//   const [products, setProducts] = useState([]);
+//   const [error, setError] = useState('');
 
-  useEffect(() => {
-    const fetchProducts = async () => {
-      try {
-        const response = await fetch('https://fakestoreapi.com/products?limit=1');
-        const data = await response.json();
-        setProducts(data);
-      } catch (error) {
-        console.error("Error fetching products:", error);
-        setError("Failed to fetch products.");
-      }
-    };
+//   useEffect(() => {
+//     const fetchProducts = async () => {
+//       try {
+//         const response = await fetch('https://fakestoreapi.com/products?limit=1');
+//         const data = await response.json();
+//         setProducts(data);
+//       } catch (error) {
+//         console.error("Error fetching products:", error);
+//         setError("Failed to fetch products.");
+//       }
+//     };
 
-    fetchProducts();
-  }, []);
+//     fetchProducts();
+//   }, []);
 
-  const addToCart = (product) => {
-    // Add your add to cart logic here
-    console.log(`Added ${product.title} to cart`);
-  };
+//   const addToCart = (product) => {
+//     // Add your add to cart logic here
+//     console.log(`Added ${product.title} to cart`);
+//   };
 
-  return (
-    <div className='product-display'>
-      {error && <p>{error}</p>}
-      {products.map((product) => (
-        <div key={product.id}>
-          <h2>{product.title}</h2>
-          <img src={product.image} alt={product.title} />
-          <p>${product.price}</p>
-          <button onClick={() => addToCart(product)}>Add to Cart</button>
-        </div>
-      ))}
-    </div>
-  );
-};
+//   return (
+//     <div className='product-display'>
+//       {error && <p>{error}</p>}
+//       {products.map((product) => (
+//         <div key={product.id}>
+//           <h2>{product.title}</h2>
+//           <img src={product.image} alt={product.title} />
+//           <p>${product.price}</p>
+//           <button onClick={() => addToCart(product)}>Add to Cart</button>
+//         </div>
+//       ))}
+//     </div>
+//   );
+// };
 
-export default FetchProducts;
+// export default FetchProducts;
